pyproject.toml
[tool.poetry]
name = "smiles-rl"
version = "2.0.0"
description = ""
authors = ["Hampus Gummesson Svensson <hampusgs@gmail.com>"]
readme = "README.md"
packages = [{include = "smiles_rl"}]

[tool.poetry.dependencies]
python = ">= 3.9, < 3.10"
reinvent-scoring = "^0.0.73"
reinvent-chemistry = "^0.0.51"
numpy = "^1.24.1"
dacite = "^1.7.0"
pandas = "^2.1.4"
pydantic = "^2.6.3"
requests = "^2.28.1"
rdkit = "^2023.9.5"
pathos = "^0.3.0"
scipy = "^1.10.0"
tensorboard = "^2.11.0"
scikit-learn = "1.2.2"
reinvent-models = "^0.0.14"
PyTDC = "^0.4.17"
openeye-toolkits = {version = "^2024.1.0", source = "openeye-toolkits-src"}
torch = {version = "2.2.1+cu118", source = "pytorch-gpu-src"}

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pytest = "^7.2.0"


[[tool.poetry.source]]
name = "openeye-toolkits-src"
url = "https://pypi.anaconda.org/OpenEye/simple"
priority = "explicit"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
